feat: Implement IProcessEx interface and enhance pseudo-terminal functionality with platform-specific methods
